requirements.txt
mistralai
python-dotenv
sentence-transformers
faiss-cpu
streamlit